+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1.common/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/APIRevisionDTO.java
@@ -21,13 +21,14 @@ import javax.validation.Valid;
 
 public class APIRevisionDTO   {
   
-    private String id = null;
+    private Integer id = null;
+    private String uuid = null;
     private String description = null;
     private APIRevisionAPIInfoDTO apiInfo = null;
 
   /**
    **/
-  public APIRevisionDTO id(String id) {
+  public APIRevisionDTO id(Integer id) {
     this.id = id;
     return this;
   }
@@ -35,11 +36,28 @@ public class APIRevisionDTO   {
   
   @ApiModelProperty(example = "1", value = "")
   @JsonProperty("id")
-  public String getId() {
+  public Integer getId() {
     return id;
   }
-  public void setId(String id) {
+  public void setId(Integer id) {
     this.id = id;
+  }
+
+  /**
+   **/
+  public APIRevisionDTO uuid(String uuid) {
+    this.uuid = uuid;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "c26b2b9b-4632-4ca4-b6f3-521c8863990c", value = "")
+  @JsonProperty("uuid")
+  public String getUuid() {
+    return uuid;
+  }
+  public void setUuid(String uuid) {
+    this.uuid = uuid;
   }
 
   /**
@@ -88,13 +106,14 @@ public class APIRevisionDTO   {
     }
     APIRevisionDTO apIRevision = (APIRevisionDTO) o;
     return Objects.equals(id, apIRevision.id) &&
+        Objects.equals(uuid, apIRevision.uuid) &&
         Objects.equals(description, apIRevision.description) &&
         Objects.equals(apiInfo, apIRevision.apiInfo);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(id, description, apiInfo);
+    return Objects.hash(id, uuid, description, apiInfo);
   }
 
   @Override
@@ -103,6 +122,7 @@ public class APIRevisionDTO   {
     sb.append("class APIRevisionDTO {\n");
     
     sb.append("    id: ").append(toIndentedString(id)).append("\n");
+    sb.append("    uuid: ").append(toIndentedString(uuid)).append("\n");
     sb.append("    description: ").append(toIndentedString(description)).append("\n");
     sb.append("    apiInfo: ").append(toIndentedString(apiInfo)).append("\n");
     sb.append("}");
