+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/.openapi-generator/publisher-api.yaml.sha256
@@ -1,2 +1,1 @@
 673c597faceb2e8c6c2df26da1f563de57a77d3f8074e85370e6b31bd61b5f01
-
