--- components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/APIRevisionAPIInfoDTO.java
@@ -1,221 +1,0 @@
-package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import javax.validation.constraints.*;
-
-
-import io.swagger.annotations.*;
-import java.util.Objects;
-
-import javax.xml.bind.annotation.*;
-import org.wso2.carbon.apimgt.rest.api.util.annotations.Scope;
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-import javax.validation.Valid;
-
-
-
-public class APIRevisionAPIInfoDTO   {
-  
-    private String id = null;
-    private String name = null;
-    private String description = null;
-    private String context = null;
-    private String version = null;
-    private String provider = null;
-    private String type = null;
-    private String lifeCycleStatus = null;
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO id(String id) {
-    this.id = id;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "01234567-0123-0123-0123-012345678901", value = "")
-  @JsonProperty("id")
-  public String getId() {
-    return id;
-  }
-  public void setId(String id) {
-    this.id = id;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO name(String name) {
-    this.name = name;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "CalculatorAPI", value = "")
-  @JsonProperty("name")
-  public String getName() {
-    return name;
-  }
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO description(String description) {
-    this.description = description;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "A calculator API that supports basic operations", value = "")
-  @JsonProperty("description")
-  public String getDescription() {
-    return description;
-  }
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO context(String context) {
-    this.context = context;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "CalculatorAPI", value = "")
-  @JsonProperty("context")
-  public String getContext() {
-    return context;
-  }
-  public void setContext(String context) {
-    this.context = context;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO version(String version) {
-    this.version = version;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "1.0.0", value = "")
-  @JsonProperty("version")
-  public String getVersion() {
-    return version;
-  }
-  public void setVersion(String version) {
-    this.version = version;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO provider(String provider) {
-    this.provider = provider;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "admin", value = "")
-  @JsonProperty("provider")
-  public String getProvider() {
-    return provider;
-  }
-  public void setProvider(String provider) {
-    this.provider = provider;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO type(String type) {
-    this.type = type;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "HTTP", value = "")
-  @JsonProperty("type")
-  public String getType() {
-    return type;
-  }
-  public void setType(String type) {
-    this.type = type;
-  }
-
-  /**
-   **/
-  public APIRevisionAPIInfoDTO lifeCycleStatus(String lifeCycleStatus) {
-    this.lifeCycleStatus = lifeCycleStatus;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "CREATED", value = "")
-  @JsonProperty("lifeCycleStatus")
-  public String getLifeCycleStatus() {
-    return lifeCycleStatus;
-  }
-  public void setLifeCycleStatus(String lifeCycleStatus) {
-    this.lifeCycleStatus = lifeCycleStatus;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    APIRevisionAPIInfoDTO apIRevisionAPIInfo = (APIRevisionAPIInfoDTO) o;
-    return Objects.equals(id, apIRevisionAPIInfo.id) &&
-        Objects.equals(name, apIRevisionAPIInfo.name) &&
-        Objects.equals(description, apIRevisionAPIInfo.description) &&
-        Objects.equals(context, apIRevisionAPIInfo.context) &&
-        Objects.equals(version, apIRevisionAPIInfo.version) &&
-        Objects.equals(provider, apIRevisionAPIInfo.provider) &&
-        Objects.equals(type, apIRevisionAPIInfo.type) &&
-        Objects.equals(lifeCycleStatus, apIRevisionAPIInfo.lifeCycleStatus);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, name, description, context, version, provider, type, lifeCycleStatus);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class APIRevisionAPIInfoDTO {\n");
-    
-    sb.append("    id: ").append(toIndentedString(id)).append("\n");
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    description: ").append(toIndentedString(description)).append("\n");
-    sb.append("    context: ").append(toIndentedString(context)).append("\n");
-    sb.append("    version: ").append(toIndentedString(version)).append("\n");
-    sb.append("    provider: ").append(toIndentedString(provider)).append("\n");
-    sb.append("    type: ").append(toIndentedString(type)).append("\n");
-    sb.append("    lifeCycleStatus: ").append(toIndentedString(lifeCycleStatus)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
