--- components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/APIRevisionListDTO.java
@@ -1,106 +1,0 @@
-package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import java.util.ArrayList;
-import java.util.List;
-import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.APIRevisionDTO;
-import javax.validation.constraints.*;
-
-
-import io.swagger.annotations.*;
-import java.util.Objects;
-
-import javax.xml.bind.annotation.*;
-import org.wso2.carbon.apimgt.rest.api.util.annotations.Scope;
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-import javax.validation.Valid;
-
-
-
-public class APIRevisionListDTO   {
-  
-    private Integer count = null;
-    private List<APIRevisionDTO> list = new ArrayList<APIRevisionDTO>();
-
-  /**
-   * Number of API revisions returned 
-   **/
-  public APIRevisionListDTO count(Integer count) {
-    this.count = count;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "1", value = "Number of API revisions returned ")
-  @JsonProperty("count")
-  public Integer getCount() {
-    return count;
-  }
-  public void setCount(Integer count) {
-    this.count = count;
-  }
-
-  /**
-   **/
-  public APIRevisionListDTO list(List<APIRevisionDTO> list) {
-    this.list = list;
-    return this;
-  }
-
-  
-  @ApiModelProperty(value = "")
-      @Valid
-  @JsonProperty("list")
-  public List<APIRevisionDTO> getList() {
-    return list;
-  }
-  public void setList(List<APIRevisionDTO> list) {
-    this.list = list;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    APIRevisionListDTO apIRevisionList = (APIRevisionListDTO) o;
-    return Objects.equals(count, apIRevisionList.count) &&
-        Objects.equals(list, apIRevisionList.list);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(count, list);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class APIRevisionListDTO {\n");
-    
-    sb.append("    count: ").append(toIndentedString(count)).append("\n");
-    sb.append("    list: ").append(toIndentedString(list)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
